% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{now}
\alias{now}
\title{now
function to init this R environment with a pkgr.yml file and isolate it with pkgr.yml}
\usage{
now(pkgs = c())
}
\arguments{
\item{pkgs}{vector of packages to include in initial  pkgr.yml file.}
}
\description{
now
function to init this R environment with a pkgr.yml file and isolate it with pkgr.yml
}
