% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{r_pkgr_add}
\alias{r_pkgr_add}
\title{Wrapper for pkgr add}
\usage{
r_pkgr_add(exe_path, pkg)
}
\arguments{
\item{exe_path}{path to pkgr executable}

\item{pkgr}{pkg to add}
}
\description{
Wrapper for pkgr add
}
