% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{r_pkgr_install}
\alias{r_pkgr_install}
\title{Wrapper for pkgr install}
\usage{
r_pkgr_install(exe_path)
}
\arguments{
\item{exe_path}{path to pkgr executable}
}
\description{
Wrapper for pkgr install
}
