% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{r_pkgr_remove}
\alias{r_pkgr_remove}
\title{Wrapper for pkgr remove}
\usage{
r_pkgr_remove(exe_path, pkg)
}
\arguments{
\item{exe_path}{path to pkgr executable}

\item{pkg}{pkg to remove}
}
\description{
Wrapper for pkgr remove
}
